Extract system code into module
