+++ crates/scheduler/src/schedule/mod.rs
@@ -16,7 +16,8 @@ use daggy::petgraph::dot::{Config, Dot};
 use daggy::petgraph::visit::{IntoNeighbors, IntoNeighborsDirected, IntoNodeIdentifiers};
 use daggy::petgraph::{visit, Incoming};
 use daggy::{Dag, NodeIndex, WouldCycle};
-use ecs::{Schedule, ScheduleError, ScheduleResult, System, SystemExecutionGuard};
+use ecs::systems::System;
+use ecs::{Schedule, ScheduleError, ScheduleResult, SystemExecutionGuard};
 use itertools::Itertools;
 use std::fmt::{Debug, Display, Formatter};
 use thiserror::Error;
